# backend/openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI小说创作系统",
    "description": "基于多Agent协作的智能小说创作平台",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/consistency/test": {
      "get": {
        "tags": [
          "一致性检查"
        ],
        "summary": "Test Consistency Route",
        "description": "测试路由是否正常工作",
        "operationId": "test_consistency_route_api_consistency_test_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/consistency/check-stream": {
      "post": {
        "tags": [
          "一致性检查"
        ],
        "summary": "Check Consistency Stream",
        "operationId": "check_consistency_stream_api_consistency_check_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsistencyCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConsistencyCheckRequest": {
        "properties": {
          "novel_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Novel Id"
          },
          "chapter": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Chapter"
          },
          "content": {
            "type": "string",
            "maxLength": 50000,
            "minLength": 1,
            "title": "Content"
          },
          "current_day": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Current Day",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "novel_id",
          "chapter",
          "content"
        ],
        "title": "ConsistencyCheckRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}